refactor delegateNoAccessModal
