+++ src/hooks/useDelegateUserDetails.ts
@@ -1,6 +1,5 @@
 import {useOnyx} from 'react-native-onyx';
 import AccountUtils from '@libs/AccountUtils';
-import DelegateRestrictedAccessPromptText from '@libs/DelegateRestrictedAccessPromptText';
 import ONYXKEYS from '@src/ONYXKEYS';
 import useCurrentUserPersonalDetails from './useCurrentUserPersonalDetails';
 
@@ -8,11 +7,11 @@ function useDelegateUserDetails() {
     const currentUserDeatils = useCurrentUserPersonalDetails();
     const [currentUserAccountDetails] = useOnyx(ONYXKEYS.ACCOUNT);
     const isDelegateAccessRestricted = AccountUtils.isDelegateOnlySubmitter(currentUserAccountDetails);
-    const delegateNoAccessPrompt = DelegateRestrictedAccessPromptText(currentUserDeatils?.login ?? '');
+    const delegatorEmail = currentUserDeatils?.login;
 
     return {
         isDelegateAccessRestricted,
-        delegateNoAccessPrompt,
+        delegatorEmail,
     };
 }
